requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.2.0
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
cached-property==1.5.2
certifi==2019.9.11
cffi==1.13.1
chardet==4.0.0
Click==7.0
colorama==0.4.1
cryptography==2.8
defusedxml==0.6.0
distlib==0.3.1
dominate==2.4.0
elastic-apm==5.3.2
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.8
flask-restful-swagger-2==0.35
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.36.0
Flask-WTF==0.14.2
gunicorn==20.0.4
idna==3.1
importlib-metadata==1.6.1
importlib-resources==5.1.0
inflect==5.0.2
isodate==0.6.0
isort==4.3.9
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
mccabe==0.6.1
pipenv==2018.11.26
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.0
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests
requests-file==1.5.1
requests-toolbelt==0.9.1
safrs==2.10.4
six==1.12.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.10
suds-jurko==0.6
toml==0.10.2
typed-ast==1.3.1
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==16.7.5
virtualenv-clone==0.5.3
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.1
WTForms==2.2
zeep==4.0.0
zipp==3.4.0
